requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.3
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.1
Werkzeug==3.0.3
